requirements.txt
streamlit>=1.30.0
tensorflow>=2.10.0
numpy
pillow
